Separate model module.
